requirements.txt
streamlit
supabase
pandas
numpy
tensorflow
Pillow
protobuf
